netlify/functions/songs-add/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
supabase==2.4.2
